Create GeoLocation System

+++ src/controllers/auth.js
@@ -13,14 +13,15 @@ class Auth {
 			await userNotExists(req.body.email)
 			req.body.password = await encrypt(req.body.password)
 
-			await sharp(req.file.path)
+			if(req.file) {
+				await sharp(req.file.path)
 				.rotate()
 				.resize(150)
 				.toBuffer()
 				.then(buffer => { req.body.image = buffer.toString('base64') })
 
-			await fs.unlink(req.file.path, () => { })
-			
+				await fs.unlink(req.file.path, () => { })
+			}
 			const user = await User.create(req.body)
 
 			user.password = undefined
@@ -28,7 +29,6 @@ class Auth {
 
 			return res.json({ user, token })
 		} catch (err) {
-			console.log(err)
 			return res.status(400).send(err)
 		}
 	}
